Extract `AuctionResultsControls`
