+++ src/Apps/Artist/Routes/AuctionResults/Components/TableSidebar.tsx
@@ -2,16 +2,14 @@ import React from "react"
 import { Media } from "Utils/Responsive"
 
 import { Flex, Row, Separator } from "@artsy/palette"
-import { AuctionFilters } from "./Components/AuctionFilters"
+import { AuctionFilters } from "./AuctionFilters"
 
 export const TableSidebar = () => {
   return (
     <Flex flexDirection="column">
-      <Media greaterThan="xs">
-        <Row>
-          <Separator />
-        </Row>
-      </Media>
+      <Row>
+        <Separator />
+      </Row>
       <AuctionFilters />
     </Flex>
   )
